requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
diff-match-patch==20241021
Django==5.2.3
django-admin-geomap==1.6
django-allauth==65.9.0
django-appconf==1.1.0
django-autocomplete-light==3.12.1
django-ckeditor-5==0.2.18
django-import-export==4.3.7
django-model-utils==5.0.0
django-phonenumber-field==8.1.0
django-select2==8.4.1
django-webp==3.0.0
idna==3.10
jwt==1.3.1
mysqlclient==2.2.7
phonenumbers==9.0.7
pillow==11.2.1
pycparser==2.22
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.0
urllib3==2.4.0
utils==1.0.2
whitenoise==6.9.0
